pyproject.toml
[tool.poetry]
name = "yolov8-training"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.9.0"
# ultralytics = "8.0.97"
ultralytics = "^8.0.97"
scikit-learn = "^1.5.0"
albumentations = "^1.4.10"
imageio = "^2.34.2"
tabulate = "^0.9.0"
ffmpeg-python = "^0.2.0"
setuptools = "^70.3.0"
imagehash = "^4.3.1"
imgaug = "^0.4.0"
pillow = "^11.0.0"
numpy = "^2.2.0"
scikit-image = "^0.25.0"
tqdm = "^4.67.1"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
opencv-python = "^4.10.0.84"
dvc = "^3.60.1"
dvc-ssh = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
